Split constants and lookup stuff to Base Controller to use with ShiftTracks
